+++ src/m_config.c
@@ -120,7 +120,8 @@ extern dboolean         r_mirroredweapons;
 extern dboolean         r_playersprites;
 extern dboolean         r_rockettrails;
 extern dboolean         r_shadows;
-extern int              r_shakescreen;
+extern dboolean         r_shake_damage;
+extern dboolean         r_shake_explode;
 extern dboolean         r_translucency;
 extern int              s_musicvolume;
 extern dboolean         s_randommusic;
@@ -285,7 +286,8 @@ static default_t cvars[] =
     CONFIG_VARIABLE_INT          (r_rockettrails,                                    BOOLALIAS  ),
     CONFIG_VARIABLE_INT          (r_screensize,                                      NOALIAS    ),
     CONFIG_VARIABLE_INT          (r_shadows,                                         BOOLALIAS  ),
-    CONFIG_VARIABLE_INT_PERCENT  (r_shakescreen,                                     NOALIAS    ),
+    CONFIG_VARIABLE_INT          (r_shake_damage,                                    BOOLALIAS  ),
+    CONFIG_VARIABLE_INT          (r_shake_explode,                                   BOOLALIAS  ),
     CONFIG_VARIABLE_INT          (r_translucency,                                    BOOLALIAS  ),
     CONFIG_VARIABLE_INT_PERCENT  (s_musicvolume,                                     NOALIAS    ),
     CONFIG_VARIABLE_INT          (s_randommusic,                                     BOOLALIAS  ),
@@ -789,7 +791,11 @@ static void M_CheckCVARs(void)
     if (r_shadows != false && r_shadows != true)
         r_shadows = r_shadows_default;
 
-    r_shakescreen = BETWEEN(r_shakescreen_min, r_shakescreen, r_shakescreen_max);
+    if (r_shake_damage != false && r_shake_damage != true)
+        r_shake_damage = r_shake_damage_default;
+
+    if (r_shake_explode != false && r_shake_explode != true)
+        r_shake_explode = r_shake_explode_default;
 
     if (r_translucency != false && r_translucency != true)
         r_translucency = r_translucency_default;